# app/Http/Controllers/UserApi/SubscriptionController.php
<?php

namespace App\Http\Controllers\UserApi;

use App\Http\Controllers\Controller;
use App\Models\Superstar;
use App\Models\UserGoogle;
use Illuminate\Http\Request;
use Illuminate\Http\Response;

class SubscriptionController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/user/subscriptions",
     *     summary="Get user subscriptions",
     *     description="Get authenticated user's subscriptions with pagination. Returns superstars with their details",
     *     tags={"User Subscriptions"},
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="per_page",
     *         in="query",
     *         description="Number of items per page",
     *         required=false,
     *         @OA\Schema(type="integer", default=15)
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Subscriptions retrieved successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="data", type="array", @OA\Items(type="object")),
     *             @OA\Property(property="pagination", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function index(Request $request)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        $perPage = $request->get('per_page', 15);
        
        $subscriptions = $user->superstars()
            ->select([
                'superstars.id',
                'superstars.user_id',
                'superstars.display_name',
                'superstars.bio',
                'superstars.price_per_minute',
                'superstars.is_available',
                'superstars.rating',
                'superstars.total_followers',
                'superstars.status',
                'subscribes.created_at as subscribed_at'
            ])
            ->orderBy('subscribes.created_at', 'desc')
            ->paginate($perPage);

        return response()->json([
            'success' => true,
            'data' => $subscriptions,
        ], Response::HTTP_OK);
    }

    /**
     * @OA\Post(
     *     path="/api/user/subscriptions",
     *     summary="Subscribe to superstar",
     *     description="Subscribe to a superstar and increases total_followers count",
     *     tags={"User Subscriptions"},
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"superstar_id"},
     *             @OA\Property(property="superstar_id", type="integer", example=1)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Subscription successful",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Subscribed successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function store(Request $request)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        $validated = $request->validate([
            'superstar_id' => 'required|exists:superstars,id',
        ]);

        $superstar = Superstar::findOrFail($validated['superstar_id']);

        // Check if already subscribed
        if ($user->superstars()->where('superstar_id', $superstar->id)->exists()) {
            return response()->json([
                'success' => false,
                'message' => 'You are already subscribed to this superstar',
            ], Response::HTTP_BAD_REQUEST);
        }

        // Subscribe
        $user->superstars()->attach($superstar->id);

        // Increase total_followers
        $superstar->increment('total_followers');

        return response()->json([
            'success' => true,
            'message' => 'Successfully subscribed to superstar',
            'data' => [
                'superstar' => [
                    'id' => $superstar->id,
                    'display_name' => $superstar->display_name,
                    'total_followers' => $superstar->total_followers,
                ],
            ],
        ], Response::HTTP_CREATED);
    }

    /**
     * @OA\Delete(
     *     path="/api/user/subscriptions/{superstarId}",
     *     summary="Unsubscribe from superstar",
     *     description="Unsubscribe from a superstar and decreases total_followers count",
     *     tags={"User Subscriptions"},
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="superstarId",
     *         in="path",
     *         description="Superstar ID to unsubscribe from",
     *         required=true,
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Unsubscription successful",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Unsubscribed successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Subscription not found"
     *     )
     * )
     */
    public function destroy($superstarId)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        $superstar = Superstar::findOrFail($superstarId);

        // Check if subscribed
        if (!$user->superstars()->where('superstar_id', $superstar->id)->exists()) {
            return response()->json([
                'success' => false,
                'message' => 'You are not subscribed to this superstar',
            ], Response::HTTP_BAD_REQUEST);
        }

        // Unsubscribe
        $user->superstars()->detach($superstar->id);

        // Decrease total_followers (but don't go below 0)
        if ($superstar->total_followers > 0) {
            $superstar->decrement('total_followers');
        }

        return response()->json([
            'success' => true,
            'message' => 'Successfully unsubscribed from superstar',
            'data' => [
                'superstar' => [
                    'id' => $superstar->id,
                    'display_name' => $superstar->display_name,
                    'total_followers' => $superstar->fresh()->total_followers,
                ],
            ],
        ], Response::HTTP_OK);
    }

    /**
     * View superstar details by ID
     * Returns all superstar details
     */
    public function show($id)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        $superstar = Superstar::select([
            'id',
            'user_id',
            'display_name',
            'bio',
            'price_per_minute',
            'is_available',
            'rating',
            'total_followers',
            'status',
        ])->findOrFail($id);

        // Check if user is subscribed
        $isSubscribed = $user->superstars()->where('superstar_id', $id)->exists();

        return response()->json([
            'success' => true,
            'data' => [
                'superstar' => $superstar,
                'is_subscribed' => $isSubscribed,
            ],
        ], Response::HTTP_OK);
    }
}