# app/Http/Controllers/UserApi/UserApiLogin.php
<?php

namespace App\Http\Controllers\UserApi;

use App\Http\Controllers\Controller;
use App\Models\UserGoogle;
use Illuminate\Http\Request;
use Illuminate\Http\Response;

class UserApiLogin extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/user/google-login",
     *     summary="Google OAuth Login",
     *     description="This endpoint handles Google OAuth login by checking if the email exists, updating or creating user data, and returning user data with Sanctum token",
     *     tags={"User Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email","username"},
     *             @OA\Property(property="email", type="string", format="email", example="user@example.com"),
     *             @OA\Property(property="username", type="string", example="johndoe"),
     *             @OA\Property(property="image", type="string", nullable=true, example="https://example.com/avatar.jpg")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Login successful",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Login successful"),
     *             @OA\Property(property="user", type="object"),
     *             @OA\Property(property="token", type="string")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function googleLogin(Request $request)
    {
        // Validate incoming request
        $validated = $request->validate([
            'email' => 'required|email',
            'username' => 'required|string|max:255',
            'image' => 'nullable|string',
        ]);

        try {
            // Check if user with this email already exists
            $user = UserGoogle::where('email', $validated['email'])->first();

            if ($user) {
                // User exists - update their information
                $user->update([
                    'username' => $validated['username'],
                    'image' => $validated['image'] ?? $user->image,
                ]);

                $message = 'Login successful';
            } else {
                // User doesn't exist - create new user
                $user = UserGoogle::create([
                    'email' => $validated['email'],
                    'username' => $validated['username'],
                    'image' => $validated['image'],
                ]);

                $message = 'Account created successfully';
            }

            // Generate Sanctum token
            $token = $user->createToken('google-auth-token')->plainTextToken;

            return response()->json([
                'success' => true,
                'message' => $message,
                'data' => [
                    'user' => [
                        'id' => $user->id,
                        'email' => $user->email,
                        'username' => $user->username,
                        'image' => $user->image,
                    ],
                    'token' => $token,
                    'token_type' => 'Bearer',
                ],
            ], Response::HTTP_OK);

        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'An error occurred during login',
                'error' => $e->getMessage(),
            ], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }

    /**
     * Get current authenticated user
     */
    public function getAuthUser(Request $request)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        return response()->json([
            'success' => true,
            'data' => [
                'id' => $user->id,
                'email' => $user->email,
                'username' => $user->username,
                'image' => $user->image,
            ],
        ], Response::HTTP_OK);
    }

    /**
     * Logout - Revoke all tokens
     */
    public function logout(Request $request)
    {
        /** @var UserGoogle|null $user */
        $user = auth('sanctum')->user();

        if (!$user || !($user instanceof UserGoogle)) {
            return response()->json([
                'success' => false,
                'message' => 'Unauthorized',
            ], Response::HTTP_UNAUTHORIZED);
        }

        // Delete all tokens for the user
        $user->tokens()->delete();

        return response()->json([
            'success' => true,
            'message' => 'Logged out successfully',
        ], Response::HTTP_OK);
    }
}