# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Rashid Backend API Documentation
  description: Complete API Documentation for Rashid Backend Application
  version: 1.0.0
  contact:
    email: admin@rashid.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://127.0.0.1:8000
    description: Rashid Backend API Server
...
paths:
...
  /api/user/google-login:
    post:
      tags:
        - User Authentication
      summary: Google OAuth Login
      description: This endpoint handles Google OAuth login by checking if the email exists, updating or creating user data, and returning user data with Sanctum token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                username:
                  type: string
                  example: johndoe
                image:
                  type: string
                  nullable: true
                  example: https://example.com/avatar.jpg
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    type: object
                  token:
                    type: string
        '422':
          description: Validation error

  # User Subscriptions
  /api/user/subscriptions:
    get:
      tags:
        - User Subscriptions
      summary: Get user subscriptions
      description: Get authenticated user's subscriptions with pagination. Returns superstars with their details
      security:
        - sanctum: []
      parameters:
        - name: per_page
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    type: object
        '401':
          description: Unauthorized

    post:
      tags:
        - User Subscriptions
      summary: Subscribe to superstar
      description: Subscribe to a superstar and increases total_followers count
      security:
        - sanctum: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - superstar_id
              properties:
                superstar_id:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Subscription successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Subscribed successfully
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /api/user/subscriptions/{superstarId}:
    delete:
      tags:
        - User Subscriptions
      summary: Unsubscribe from superstar
      description: Unsubscribe from a superstar and decreases total_followers count
      security:
        - sanctum: []
      parameters:
        - name: superstarId
          in: path
          description: Superstar ID to unsubscribe from
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Unsubscription successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Unsubscribed successfully
        '401':
          description: Unauthorized
        '404':
          description: Subscription not found

  # User Chat
...
components:
  securitySchemes:
    sanctum:
      type: apiKey
      in: header
      name: Authorization
      description: Enter token in format (Bearer <token>)